Create subpalindromes.cpp
A string is called a palindrome if it reads the same from left to right as from right to left. For example, the strings "abba", "kek" are palindromes.

A string is given. Its substring is a non-empty sequence of consecutive characters. Write a program that determines how many substrings of this string are palindromes.

A simple solution idea. Let's count the number of palindromes with odd center and then with even center. For this purpose, if the center is odd, we fix it and try to find the largest palindrome with this center by binary search. For checking we use hashes. If they are equal, the rows are equal (this is their main idea). Since a string of length 1 is always a palindrome, we will always add 1 for this case. For even centers we will do the same except adding 1. Asymptotics O(n * log(n))

There is also a faster solution for O(n) - the Manaker algorithm