Create trapping_rain_water.cpp
Given n non-negative integers representing an elevation map where the width of each bar is 1, compute how much water it can trap after raining.

Solution O(n)
The explanation: The water won't rise above the maximum, let's find it. Next, let's divide our problem by 2. Finding the amount of water before the maximum and after the maximum.
In order to solve the problem, let's understand when water is added. It happens if index < index of maximum and there is some index < i < index of maximum and a[index] < a[i]. If a[i] > a[index]. Then let's move our index to i. No water is added in this case.